refactor to custom hook
