+++ frontend/src/scenes/events/EventsTable.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useMemo } from 'react'
+import React, { useMemo } from 'react'
 import { useActions, useValues } from 'kea'
 import { EventDetails } from 'scenes/events/EventDetails'
 import { DownloadOutlined } from '@ant-design/icons'
@@ -32,6 +32,7 @@ import { LemonSwitch } from 'lib/components/LemonSwitch/LemonSwitch'
 import { propertyFilterLogic } from 'lib/components/PropertyFilters/propertyFilterLogic'
 import { teamLogic } from 'scenes/teamLogic'
 import { createActionFromEvent } from './createActionFromEvent'
+import { usePageVisibility } from 'lib/hooks/usePageVisibility'
 
 export interface FixedFilters {
     action_id?: ActionType['id']
@@ -80,31 +81,7 @@ export function EventsTable({
 
     const showLinkToPerson = !fixedFilters?.person_id
 
-    useEffect(() => {
-        // adapted from https://developer.mozilla.org/en-US/docs/Web/API/Page_Visibility_API#example
-        // Opera 12.10 and Firefox 18 and later support
-        let hidden = 'hidden'
-        let visibilityChange = 'visibilitychange'
-        // @ts-ignore - to avoid complaint that msHidden isn't on document
-        if (typeof document.msHidden !== 'undefined') {
-            hidden = 'msHidden'
-            visibilityChange = 'msvisibilitychange'
-            // @ts-ignore - to avoid complaint that webkitHidden isn't on document
-        } else if (typeof document.webkitHidden !== 'undefined') {
-            hidden = 'webkitHidden'
-            visibilityChange = 'webkitvisibilitychange'
-        }
-
-        const onVisibilityChange = (): void => {
-            setPollingActive(!document[hidden])
-        }
-
-        document.addEventListener(visibilityChange, onVisibilityChange)
-
-        return function cleanUp() {
-            document.removeEventListener(visibilityChange, onVisibilityChange)
-        }
-    }, [])
+    usePageVisibility(setPollingActive)
 
     const newEventsRender = (
         { date_break, new_events }: EventsTableRowItem,
